test: Test Clock
