+++ core/test/clock/Clock.api.js
@@ -1,0 +1,11 @@
+module.exports = [
+    'setScale',
+    'getScale',
+    'step',
+    'now',
+    'getTime',
+    'getFrame',
+    'setTimeout',
+    'setInterval',
+    'clearTimer'
+];
